Fix max allowed sandbox time

+++ packages/api/internal/cache/instance/instance.go
@@ -39,13 +39,13 @@ type InstanceCache struct {
 
 	logger *zap.SugaredLogger
 
-	counter   metric.Int64UpDownCounter
-	analytics analyticscollector.AnalyticsCollectorClient
+	counter   *metric.Int64UpDownCounter
+	analytics *analyticscollector.AnalyticsCollectorClient
 
 	mu sync.Mutex
 }
 
-func NewCache(analytics analyticscollector.AnalyticsCollectorClient, logger *zap.SugaredLogger, deleteInstance func(data InstanceInfo, purge bool) *api.APIError, initialInstances []*InstanceInfo, counter metric.Int64UpDownCounter) *InstanceCache {
+func NewCache(analytics *analyticscollector.AnalyticsCollectorClient, logger *zap.SugaredLogger, deleteInstance func(data InstanceInfo, purge bool) *api.APIError, initialInstances []*InstanceInfo, counter *metric.Int64UpDownCounter) *InstanceCache {
 	// We will need to either use Redis or Consul's KV for storing active sandboxes to keep everything in sync,
 	// right now we load them from Orchestrator
 	cache := ttlcache.New(
@@ -63,23 +63,28 @@ func NewCache(analytics analyticscollector.AnalyticsCollectorClient, logger *zap
 
 	cache.OnInsertion(func(ctx context.Context, i *ttlcache.Item[string, InstanceInfo]) {
 		instanceInfo := i.Value()
-		_, err := analytics.InstanceStarted(ctx, &analyticscollector.InstanceStartedEvent{
-			InstanceId:    instanceInfo.Instance.SandboxID,
-			EnvironmentId: instanceInfo.Instance.TemplateID,
-			BuildId:       instanceInfo.BuildID.String(),
-			TeamId:        instanceInfo.TeamID.String(),
-			Timestamp:     timestamppb.Now(),
-		})
-		if err != nil {
-			errMsg := fmt.Errorf("error when sending analytics event: %w", err)
-			telemetry.ReportCriticalError(ctx, errMsg)
+		if analytics != nil {
+			_, err := (*analytics).InstanceStarted(ctx, &analyticscollector.InstanceStartedEvent{
+				InstanceId:    instanceInfo.Instance.SandboxID,
+				EnvironmentId: instanceInfo.Instance.TemplateID,
+				BuildId:       instanceInfo.BuildID.String(),
+				TeamId:        instanceInfo.TeamID.String(),
+				Timestamp:     timestamppb.Now(),
+			})
+			if err != nil {
+				errMsg := fmt.Errorf("error when sending analytics event: %w", err)
+				telemetry.ReportCriticalError(ctx, errMsg)
+			}
 		}
 	})
+
 	cache.OnEviction(func(ctx context.Context, er ttlcache.EvictionReason, i *ttlcache.Item[string, InstanceInfo]) {
 		if er == ttlcache.EvictionReasonExpired || er == ttlcache.EvictionReasonDeleted {
-			err := deleteInstance(i.Value(), true)
-			if err != nil {
-				logger.Errorf("Error deleting instance (%v)\n: %v", er, err.Err)
+			if deleteInstance != nil {
+				err := deleteInstance(i.Value(), true)
+				if err != nil {
+					logger.Errorf("Error deleting instance (%v)\n: %v", er, err.Err)
+				}
 			}
 
 			instanceCache.UpdateCounter(i.Value(), -1)
